TuxBot/requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
draccus==0.11.5
fastapi==0.117.1
filelock==3.19.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.0
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mergedeep==1.3.4
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
packaging==25.0
peft==0.17.1
psutil==7.1.0
pydantic==2.11.9
pydantic_core==2.33.2
PyYAML==6.0.2
pyyaml-include==1.4.1
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
setuptools==78.1.1
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
